rop-yasm-8cc, rop-ps4-8cc: enable linking to hand-written ROP
ABI:
Symbol names need to start with _ to be visible from C.
Symbol names starting with L must not be used.
Emulated stack is in rdi, the first qword is the return address for pop rsp.
Then arguments follow in CDECL-like fashion, one qword per argument.
In addition to normal x86_64 ABI:
* r8 is callee-preserved
* rdi on exit must be 8 bytes higher than on entry (return address is popped)